newer wget to fetch source

+++ all/iproute2-3.17.0-2/Fetch-source.sh
@@ -3,7 +3,6 @@
 SRC=iproute2-3.17.0.tar.gz
 DST=/var/spool/src/$SRC
 
-pkg_install wget-1.12-2 || exit 2
+pkg_install wget-1.18-1 || exit 2
 
-[ -s "$DST" ] || wget -O $DST http://ftp.sunet.se/pub/Linux/distributions/bifrost/download/src/$SRC || \
- wget --no-check-certificate -O $DST https://www.kernel.org/pub/linux/utils/net/iproute2/$SRC
+[ -s "$DST" ] || wget --no-check-certificate -O $DST https://www.kernel.org/pub/linux/utils/net/iproute2/$SRC
